Win64 build
R=rossberg@chromium.org,mstarzinger@chromium.org

Review URL: https://codereview.chromium.org/14392013

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@14440 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ test/cctest/test-api.cc
@@ -2263,7 +2263,7 @@ THREADED_TEST(ArrayBuffer) {
   Local<v8::ArrayBuffer> ab = v8::ArrayBuffer::New(1024);
   HEAP->CollectAllGarbage(i::Heap::kNoGCFlags);
 
-  CHECK_EQ(1024, ab->ByteLength());
+  CHECK_EQ(1024, static_cast<int>(ab->ByteLength()));
   uint8_t* data = static_cast<uint8_t*>(ab->Data());
   ASSERT(data != NULL);
   env->Global()->Set(v8_str("ab"), ab);
@@ -2288,7 +2288,7 @@ THREADED_TEST(ArrayBuffer) {
                       "u8_a[0] = 0xAA;"
                       "u8_a[1] = 0xFF; u8_a.buffer");
   Local<v8::ArrayBuffer> ab1 = v8::ArrayBuffer::Cast(*result);
-  CHECK_EQ(2, ab1->ByteLength());
+  CHECK_EQ(2, static_cast<int>(ab1->ByteLength()));
   uint8_t* ab1_data = static_cast<uint8_t*>(ab1->Data());
   CHECK_EQ(0xAA, ab1_data[0]);
   CHECK_EQ(0xFF, ab1_data[1]);
@@ -2300,7 +2300,7 @@ THREADED_TEST(ArrayBuffer) {
   uint8_t* my_data = new uint8_t[100];
   memset(my_data, 0, 100);
   Local<v8::ArrayBuffer> ab3 = v8::ArrayBuffer::New(my_data, 100);
-  CHECK_EQ(100, ab3->ByteLength());
+  CHECK_EQ(100, static_cast<int>(ab3->ByteLength()));
   CHECK_EQ(my_data, ab3->Data());
   env->Global()->Set(v8_str("ab3"), ab3);
   result = CompileRun("var u8_b = new __Uint8Array(ab3);"
